// src/store/brandStore/action.enum.ts
enum actionTypes {
  GET_BRANDS_SUCCESS = "GET_BRANDS_SUCCESS",
  GET_BRANDS_PENDING = "GET_BRANDS_PENDING",
  GET_BRANDS_FAILED = "GET_BRANDS_FAILED",

  ADD_BRAND_SUCCESS = "ADD_BRAND_SUCCESS",
  ADD_BRAND_FAILED = "ADD_BRAND_FAILED",
  ADD_BRAND_PENDING = "ADD_BRAND_PENDING",

  DELETE_BRAND_SUCCESS = "DELETE_BRAND_SUCCESS",
  DELETE_BRAND_FAILED = "DELETE_BRAND_FAILED",
  DELETE_BRAND_PENDING = "DELETE_BRAND_PENDING",
}

export default actionTypes;